+++ src/Entities/Transaction.php
@@ -86,6 +86,18 @@ final class Transaction extends Entity
     }
 
     /**
+     * Can be specified only if transaction was processed outside (isProcessedOutside = true).
+     *
+     * @param $value
+     *
+     * @return $this
+     */
+    public function setProcessedTime($value)
+    {
+        return $this->setAttribute('processedTime', $value);
+    }
+
+    /**
      * @return string
      */
     public function getScheduledTime()
@@ -569,5 +581,39 @@ final class Transaction extends Entity
     public function getPurchaseCurrency()
     {
         return $this->getAttribute('purchaseCurrency');
+    }
+
+    /**
+     * @return string
+     */
+    public function getReportAmount()
+    {
+        return $this->getAttribute('reportAmount');
+    }
+
+    /**
+     * @return string
+     */
+    public function getReportCurrency()
+    {
+        return $this->getAttribute('reportCurrency');
+    }
+
+    /**
+     * @return bool
+     */
+    public function getIsProcessedOutside()
+    {
+        return $this->getAttribute('isProcessedOutside');
+    }
+
+    /**
+     * @param bool $value
+     *
+     * @return $this
+     */
+    public function setIsProcessedOutside($value)
+    {
+        return $this->setAttribute('isProcessedOutside', $value);
     }
 }