+++ CHANGELOG.md
@@ -21,12 +21,17 @@ Security - in case of vulnerabilities.
 - [x] Added new property to `Customer`: `websiteId`
 - [x] Added new property to `Transaction`: `requestId`
 - [x] Added new property to `Website`: `organizationId`
-- [x] Added `AlternativeInstrument`, `Customer Timeline` 
+- [x] Added `CheckInstrument`, `AlternativeInstrument`, `Customer Timeline` 
 - [x] Added `iDEAL`, `Klarna`, `Interac`, `Onlineueberweisen` to supported alternative payment instruments
 - [x] Added `description` field to `ValuesList` 
 - [x] Added `reissue` method to `InvoiceService` 
 - [x] Added `dueTime` to `InvoiceService::issue()` 
-- [x] Added new properties to `Transaction`: `requestAmount`, `requestCurrency`, `purchaseAmount`, `purchaseCurrency`
+- [x] Added new properties to `Transaction`: `requestAmount`, `requestCurrency`, `purchaseAmount`, `purchaseCurrency`, `reportAmount`, `reportCurrency`, `isProcessedOutside`
+- [x] Added field `additionalCriteria` into `GatewayAccount` object
+- [x] Added field `dueTimeShift` into `BillingAnchor` object
+- [x] Renamed field `billingAnchor` into `invoiceTimeShift` in `Subscription` object
+- [x] Added `getUpcomingInvoices` and `issueUpcomingInvoice` methods to **SubscriptionService**
+- [x] Added `recalculate` method to **InvoiceService**
 
 ### Deprecated
 - [x] Deprecated `Transaction` entity method: `getPaymentCardId`
@@ -35,6 +40,7 @@ Security - in case of vulnerabilities.
 - [x] Removed Payments Queue API
 - [X] Removed `Transaction` entity method: `getPayment`
 - [x] Removed `ValuesList` field: `name` 
+- [x] Removed field `websites` from `GatewayAccount` object 
 
 ## [2.5.0] 2019-07-19